requirements.txt
flask
flask-socketio
scikit-learn
pandas
numpy
joblib
openpyxl
python-dotenv
twilio
